+++ akshare/__init__.py
@@ -2152,9 +2152,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.7.32 fix: fix stock_a_ttm_lyr interface
 1.7.33 fix: fix stock_zh_ah_daily interface
 1.7.34 fix: fix bond_cb_jsl interface
+1.7.35 fix: fix stock_market_fund_flow interface
 """
 
-__version__ = "1.7.34"
+__version__ = "1.7.35"
 __author__ = "AKFamily"
 
 import sys